fix: reject malformed workPropagation objects during validation
Preserve authored empty workPropagation objects through API mapping so
domain validation fails with a workstation-targeted error instead of
silently treating them as omitted policy. Require mode in the OpenAPI
contract and consolidate work-propagation tests to satisfy pkg-file-count.
