+++ api/components/schemas/data-models/WorkPropagation.yaml
@@ -1,5 +1,7 @@
 type: object
 additionalProperties: false
+required:
+  - mode
 description: >-
   Optional workstation policy for how downstream work receives payload content
   after this workstation completes. When omitted, downstream work uses the
